ai-service/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
openai==1.65.5